cast token_type_ids to Long

+++ model.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 from transformers import AutoModel, logging
 
@@ -28,7 +29,7 @@ class BertMetaLearning(nn.Module):
         outputs = self.model(
             data["input_ids"],
             attention_mask=data["attention_mask"],
-            token_type_ids=data["token_type_ids"],
+            token_type_ids=data["token_type_ids"].to(torch.Long),
         )
 
         features = outputs[1]  # [n, 768]
